Move grouping properties to TableView.Grouping.cs and fix XML docs
